refactor(mir): carve the declared-release carrier pass out of lowering
`lower/mod.rs` crossed its line ceiling. The shell-retirement pass and its
parent resolver are one coherent concern with no other caller, so they move
to a sibling module unchanged.
